==> src/3_plot_tone_extraction.ipynb <==
# 1. IMPORTS

## 1.1 Libraries

import pandas as pd
import numpy as np
import re
import matplotlib.pyplot as plt
from datetime import datetime
from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer

## 1.2 Data

#import merged dataset
folder = "../temporary/"
df_treated=pd.read_csv(folder + "merged_data.csv",index_col = 'Wikipedia movie ID')


# 2. PLOT TONE SCORE

# For this, we use the "VADER sentiment analysis" library which contains a dictionary. Each of the words inside are associated to a polarity score depending on how positive or negative they are. One total 
#score is computed per summary, and then it is normalized and takes a value between -1 and +1 to categorize the text as positive, negative or neutral. 

analyzer = SentimentIntensityAnalyzer()

def get_sentiment(text):
    score = analyzer.polarity_scores(text)
    compound_score = score['compound']
    if compound_score >= 0.1:
        sentiment = 'positive'
    elif compound_score <= -0.1:
        sentiment = 'negative'
    else:
        sentiment = 'neutral'
    return sentiment, compound_score

df_treated[['Sentiment', 'Compound Score']] = df_treated.apply(
    lambda row: pd.Series(get_sentiment(row['Summary'])) if pd.notna(row['Summary']) 
    else pd.Series([np.nan, np.nan]),
    axis=1)

df_treated.head(10)

# 3. EXPORT CLEAN DATASETS

df_treated.to_csv('../temporary/with_plottone_data.csv',index = True)

df_treated.isna().sum()
