==> src/4_themes_extraction.ipynb <==
# THIS NOTEBOOK WON'T RUN AS IT USES THE OPENAI API AND PRIVATE KEY.

# 1. IMPORTS

import pandas as pd
import numpy as np
import re
import matplotlib.pyplot as plt
from datetime import datetime
import openai
import os
import time


folder = "../temporary/"
df_movies=pd.read_csv(folder + "with_plottone_data.csv",index_col = 'Wikipedia movie ID')

# 2. THEME EXTRACTION

#We decided to use a ChatGPT API which allows us to extract one single theme from each movie summary. The prompt we give to the algorithm to extract the theme is : 
#"Label the movie's theme with a single word based on its plot : ". Then, we clean the results to get rid of all the unwanted characters and store everything in the movie dataframe.

#KEY is secret :-)

def chatgpt(message):
    #time.sleep(0.1)
    completion = openai.chat.completions.create(model="gpt-4o-mini",messages=[{"role": "user","content": message,},],)
    print(completion.choices[0].message.content)
    return completion.choices[0].message.content


#We are limited to 10k api requests per day -> SPLIT to train over 10 days
dfs = np.array_split(df_movies, 10)

df1, df2, df3, df4, df5, df6, df7, df8, df9, df10 = dfs

#prompt to add before each movie summary
pre = "Label the movie's theme with a single word based on its plot : "
df10['Theme'] = df10['Summary'].apply(lambda x: chatgpt(pre + x) if pd.notna(x) else np.nan)
df10['Theme'] = df10['Theme'].astype('object')


# 3. EXPORT TEMPORARY DATASETS

df10.to_csv('../temporary/df10.csv')

# 4. CLEAN THEMES

#import and concat all 10 temporary files
directory = "../temporary"

dataframes = {}

for i in range(1, 11):
    file_path = os.path.join(directory, f"df{i}.csv")
    if os.path.exists(file_path):
        dataframes[f"df{i}"] = pd.read_csv(file_path)

df_movies = pd.concat(dataframes.values(), keys=dataframes.keys(), axis=0)
df_movies.set_index('Wikipedia movie ID', inplace=True)       

df_movies

#verification if index is unique
df_movies.index.is_unique

#clean the Theme column removing unwanted caracters and bad answers

df_movies['Theme'] = df_movies['Theme'].str.replace(r'["*:.]', '', regex=True)
df_movies['Theme'] = df_movies['Theme'].str.lstrip('theme ')
df_movies['Theme'] = df_movies['Theme'].str.lstrip('Theme ')
df_movies['Theme'] = df_movies['Theme'].str.lstrip('theme of the movie can be labeled as ')
df_movies['Theme'] = df_movies['Theme'].apply(lambda x: np.nan if isinstance(x, str) and len(x) > 20 else x)

#check the most popular themes
df_movies['Theme'].value_counts()

# 5. EXPORT

#df_movies.to_csv("../generated/final_movie_dataset.csv")

# 6. ADDITIONNAL THEME CORRECTIONS

final_movie_dataset = pd.read_csv("../generated/final_movie_dataset.csv")

# Make all themes lowercase
final_movie_dataset['Theme'] = final_movie_dataset['Theme'].str.lower()

# Additionnal theme corrections for missing 't' at the beginning of the word + merging of similar themes
corrections = { 
    "ransformation": "transformation",
    "ransition": "transition",
    "ptation": "temptation",
    "ragedy": "tragedy",
    "ranscendence": "transcendence",
    "radition": "tradition",
    "ransport": "transport",
    "ransgression": "transgression",
    "ransmutation": "transmutation",
    "ransparency": "transparency",
    "ransportation": "transportation",
    "ransmigration": "transmigration",
    "rture": "torture",
    "rritory": "territory",
    "rafficking": "trafficking",
    "vriendschap": "friendship",
    "vengeance": "revenge" #grouping for analysis
}

# Function to correct themes
def correct_theme(theme):
    if pd.isna(theme):  # Handle NaN values
        return theme
    return corrections.get(theme, theme)  # Apply correction if it exists

# Apply the correction to the 'Theme' column
final_movie_dataset['Corrected_Theme'] = final_movie_dataset['Theme'].apply(correct_theme)

# Drop 'Theme' column and change 'Corrected_Theme' column to 'Theme'
final_movie_dataset = final_movie_dataset.drop(['Theme'], axis=1)
final_movie_dataset.rename({'Corrected_Theme': 'Theme'}, axis=1, inplace=True)

# 

final_movie_dataset.to_csv("../generated/final_movie_dataset.csv")
